support general-purpose build mechanism
This adds support for a general-purpose constraint building process
which can be shared across ZkC commands.
