+++ pkg/cmd/zkc/compile.go
@@ -15,27 +15,21 @@ package zkc
 import (
 	"fmt"
 	"os"
-	"path"
 
 	"github.com/consensys/go-corset/pkg/cmd/corset/debug"
-	"github.com/consensys/go-corset/pkg/ir/air"
-	"github.com/consensys/go-corset/pkg/ir/mir"
 	"github.com/consensys/go-corset/pkg/schema/register"
 	"github.com/consensys/go-corset/pkg/trace"
-	"github.com/consensys/go-corset/pkg/util"
 	"github.com/consensys/go-corset/pkg/util/field"
 	"github.com/consensys/go-corset/pkg/util/field/bls12_377"
 	"github.com/consensys/go-corset/pkg/util/field/gf251"
 	"github.com/consensys/go-corset/pkg/util/field/gf8209"
 	"github.com/consensys/go-corset/pkg/util/field/koalabear"
-	"github.com/consensys/go-corset/pkg/util/source"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast/data"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast/decl"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast/expr"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast/symbol"
 	"github.com/consensys/go-corset/pkg/zkc/compiler/ast/variable"
-	"github.com/consensys/go-corset/pkg/zkc/compiler/codegen"
 	"github.com/consensys/go-corset/pkg/zkc/constraints"
 	"github.com/consensys/go-corset/pkg/zkc/vm"
 	"github.com/consensys/go-corset/pkg/zkc/vm/instruction"
@@ -60,77 +54,17 @@ var compileCmds = []FieldAgnosticCmd{
 	{field.BLS12_377, runCompileCmd[bls12_377.Element]},
 }
 
-// BuildConfig packages up all the requirements for building artifacts.
-type BuildConfig struct {
-	// code configuration includes various things which can be turned off / on.
-	config codegen.Config
-	// field configuration
-	field field.Config
-	// metadata to include in binary output file
-	metadata []byte
-	// flags signal which layers to generate artifacts for.
-	ast, wir, fir, mir, air bool
-}
-
-// HasTarget checks whether or not at least one build target is specified.
-func (p BuildConfig) HasTarget() bool {
-	return p.ast || p.wir || p.fir || p.mir || p.air
-}
-
-// Dependencies produces a build configuration with all transitive dependencies
-// made explicit.
-func (p BuildConfig) Dependencies() BuildConfig {
-	p.mir = p.mir || p.air
-	p.fir = p.fir || p.mir
-	p.wir = p.wir || p.fir
-	p.ast = p.ast || p.wir
-	//
-	return p
-}
-
-// BuildArtifacts attempts to capture the set of build artifacts producing during a
-// compilation run.
-type BuildArtifacts[F field.Element[F]] struct {
-	// Abstract Syntax Tree
-	ast util.Option[ast.Program]
-	// Word Machine
-	wir util.Option[vm.WordMachine[vm.Uint]]
-	// Field Machine
-	fir util.Option[vm.FieldMachine[F]]
-	// MIR Constraints
-	mir util.Option[mir.Schema[F]]
-	// AIR Constraints
-	air util.Option[air.Schema[F]]
-}
-
 func runCompileCmd[F field.Element[F]](cmd *cobra.Command, args []string, field field.Config) {
 	var (
-		build  BuildConfig
+		build  = GetBuildConfig[F](cmd, field)
 		output = GetString(cmd, "output")
 	)
-	// Configure log level
-	if GetFlag(cmd, "verbose") {
-		log.SetLevel(log.DebugLevel)
-	}
-	// Configure target field
-	build.field = field
-	// Configure compiler config
-	build.config = codegen.DEFAULT_CONFIG.
-		LowerZkcNative(GetFlag(cmd, "lower-native")).
-		Vectorize(GetFlag(cmd, "vectorize")).
-		Field(field)
-	// Configure build targets
-	build.ast = GetFlag(cmd, "ast")
-	build.wir = GetFlag(cmd, "wir")
-	build.fir = GetFlag(cmd, "fir")
-	build.mir = GetFlag(cmd, "mir")
-	build.air = GetFlag(cmd, "air")
 	// Set default target (if non specified)
 	if !build.HasTarget() {
 		build.ast = true
 	}
 	// Build all artifacts
-	artifacts := Build[F](build, args...)
+	artifacts := build.Build(args...)
 	//
 	if output != "" {
 		writeArtifacts(output, build, artifacts)
@@ -140,93 +74,7 @@ func runCompileCmd[F field.Element[F]](cmd *cobra.Command, args []string, field 
 	}
 }
 
-// Build applies a build configuration with a given set of source files.
-func Build[F field.Element[F]](build BuildConfig, args ...string) BuildArtifacts[F] {
-	var (
-		errs []source.SyntaxError
-		// determine transitive dependencies
-		deps = build.Dependencies()
-		//
-		artifacts BuildArtifacts[F]
-		// Abstract Syntax Tree
-		ast ast.Program
-		// Word Machine
-		wir *vm.WordMachine[vm.Uint]
-		// Field Machine
-		fir *vm.FieldMachine[F]
-		// MIR Constraints
-		mir mir.Schema[F]
-		// AIR Constraints
-		air air.Schema[F]
-	)
-	// Check whether prebuilt binary supplied on command-line.
-	if len(args) > 0 && path.Ext(args[0]) == ".bin" {
-		// Sanity check exactly one prebuilt binary provide.
-		if len(args) != 1 {
-			log.Error("require exactly one prebuilt binary")
-			os.Exit(6)
-		} else if build.ast {
-			log.Error("cannot extract AST from prebuilt binary")
-			os.Exit(7)
-		}
-		// Single (binary) file supplied
-		wm := ReadBinaryFile[F](args[0]).WordMachine()
-		// Assign over
-		wir = &wm
-	} else {
-		// Compile source files, or print errors
-		ast = CompileSourceFiles(build.field, args...)
-		// Word-level Intermediate Representation
-		if deps.wir {
-			// Compile the AST into the top-level word machine
-			wir, errs = ast.Compile(build.config)
-			//
-			if len(errs) > 0 {
-				for _, err := range errs {
-					printSyntaxError(&err)
-				}
-				//
-				os.Exit(4)
-			}
-		}
-	}
-	// Field-level Intermediate Representation
-	if deps.fir {
-		fir = vm.LowerWordMachine[vm.Uint, F](build.field, wir)
-	}
-	// Mid-level Intermediate Representation
-	if deps.mir {
-		mir = constraints.GenerateMirConstraints(fir)
-	}
-	// Arithmetic Intermediate Representation
-	if deps.air {
-		air = constraints.GenerateAirConstraints(fir, build.field)
-	}
-	// copy over what has been requested
-	if build.ast {
-		artifacts.ast = util.Some(ast)
-	}
-	//
-	if build.wir {
-		artifacts.wir = util.Some(*wir)
-	}
-	//
-	if build.fir {
-		artifacts.fir = util.Some(*fir)
-	}
-	//
-	if build.mir {
-		artifacts.mir = util.Some(mir)
-	}
-	//
-	if build.air {
-		artifacts.air = util.Some(air)
-	}
-	//
-	return artifacts
-}
-
-func writeArtifacts[F field.Element[F]](filename string, build BuildConfig, artifacts BuildArtifacts[F]) {
+func writeArtifacts[F field.Element[F]](filename string, build BuildConfig[F], artifacts BuildArtifacts[F]) {
 	// Word-level Intermediate Representation
 	//nolint
 	if artifacts.wir.HasValue() {
@@ -551,11 +399,5 @@ func registerType(r register.Register) string {
 //nolint:errcheck
 func init() {
 	rootCmd.AddCommand(compileCmd)
-	compileCmd.Flags().Bool("ast", false, "Output Abstract Syntax Tree (AST)")
-	compileCmd.Flags().Bool("wir", false, "Output Word-level Intermediate Representation (WIR)")
-	compileCmd.Flags().Bool("fir", false, "Output Field-level Intermediate Representation (FIR)")
-	compileCmd.Flags().Bool("mir", false, "Output Mid-Level Intermediate Representation (MIR)")
-	compileCmd.Flags().Bool("air", false, "Output Arithmetic Intermediate Representation (AIR)")
 	compileCmd.Flags().StringP("output", "o", "", "specify output file for writing binary constraints")
-	compileCmd.Flags().Bool("lower-native", false, "Lower ZkC native functions into arithmetic instructions")
 }
